fix(audio): prevent integer overflow in 8-bit WAV sample conversion
Cast to int before arithmetic to avoid narrowing overflow when converting
8-bit PCM samples to 16-bit.

+++ src/core/AudioEngine.cpp
@@ -67,7 +67,7 @@ bool load(const std::string& path, AudioBuffer& out) {
     } else if (bitsPerSample == 8) {
         out.samples.resize(pcmSize);
         for (size_t i = 0; i < pcmSize; i++)
-            out.samples[i] = (static_cast<int16_t>(pcmData[i]) - 128) * 256;
+            out.samples[i] = static_cast<int16_t>((static_cast<int>(pcmData[i]) - 128) * 256);
     } else {
         Logger::warn("WAVLoader", "bitsPerSample=" + std::to_string(bitsPerSample) + " no soportado: " + path);
         return false;
